requirements.txt
gradio==4.38.1
ollama==0.2.0
psutil==5.9.8
PyPDF2==3.0.1
Pillow==10.4.0
pytesseract==0.3.10
opencv-python==4.9.0.80
PyQt6==6.7.0
plyer==2.1.0